requirements.txt
h5py
wget
torch
numpy
Pillow
matplotlib
torchvision
python-dateutil
opencv-contrib-python
Psycopg2
kafka-python
sounddevice
python-vlc
kafka-python
pika